Tests: add retries for MPI job submission check to avoid race condition
The test_slurm submits an MPI job in two nodes that will start multiple IMB-MPI1 process.
The code waits for the job to be running and then run another job
in one of the free slots of the compute instances to see if the IMB-MPI1 process is running.

We were sleeping 5 seconds before checking the MPI jobs but it can happen that the mpi job
is running but for some reasons the IMB-MPI1 process is not yet active
and this causes a failure in the tests.

The MPI jobs takes about 1m10s so we can safely have 4 retry attempts with a delay of 10 seconds.

Signed-off-by: Enrico Usai <usai@amazon.com>