Set `workaround_with_private_hostname` to true if `UseEc2Hostnames`
If and only if `UseEc2Hostnames` is set to `true`, both `use_private_hostname` and `workaround_with_private_hostname` will be set to true in Cookbook. This workaround can be turned off by setting `workaround_with_private_hostname` to `false` in `ExtraChefAttributes` on CLI.

The workaround is in Cookbook. This commit will expand the `test_cluster_in_no_internet_subnet` test to verify `srun` with `OpenMPI`, which the workaround will be able to address.

This commit will also move Slurm specific dna.json attributes to a separate function, aiming at simplifying the code

Signed-off-by: Hanwen <hanwenli@amazon.com>